requirements.txt
pyodbc
anthropic
fastapi
uvicorn
python-dotenv
